requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
asyncpg
pydantic==2.9.2  # Especifica una versión exacta de Pydantic
pydantic-settings  # Paquete adicional para BaseSettings
